Rename components/with-localized-moment to just localized-moment
There will be soon two functions exported: `withLocalizedMoment` HOC and `useLocalizedMoment` hook.
The new name is more appropriate now.
