+++ client/components/root-child/index.jsx
@@ -12,7 +12,7 @@ import { Provider as ReduxProvider } from 'react-redux';
 /**
  * Internal dependencies
  */
-import { MomentProvider } from 'components/with-localized-moment/context';
+import { MomentProvider } from 'components/localized-moment/context';
 
 export default class RootChild extends React.Component {
 	static contextTypes = {
