+++ client/lib/react-helpers/index.js
@@ -11,7 +11,7 @@ import { Provider as ReduxProvider } from 'react-redux';
 /**
  * Internal dependencies
  */
-import { MomentProvider } from 'components/with-localized-moment/context';
+import { MomentProvider } from 'components/localized-moment/context';
 
 export function concatTitle( ...parts ) {
 	return parts.join( ' › ' );
