+++ client/my-sites/activity/activity-log/index.jsx
@@ -68,7 +68,7 @@ import { requestActivityLogs } from 'state/data-getters';
 import { emptyFilter } from 'state/activity-log/reducer';
 import { isMobile } from 'lib/viewport';
 import analytics from 'lib/analytics';
-import withLocalizedMoment from 'components/with-localized-moment';
+import { withLocalizedMoment } from 'components/localized-moment';
 
 import { getPreference } from 'state/preferences/selectors';
